requirements.txt
streamlit
google-auth-oauthlib
google-api-python-client
pytube
moviepy
pyyaml
requests
Pillow
gTTS
